Promote sign-in v2 to /login, remove the v1 view
Delete the original lib/ui/login/login_view.dart (Stacked + dropdown-
based v1) and rename login_view_v2.dart in its place. Class
LoginViewV2 → LoginView. Drop the temporary /login-v2 route, its
LoginViewV2 import, and the loggingIn redirect bypass for that path.

GoRouter's /login now serves the new design directly: hero with
Register CTA, language pill, returning-user sheet with Sign in /
QRCode Sign In, server footer with Change → bottom-sheet pickers
for both language and server.
